requirements.txt
--only-binary=:all:
--no-build-isolation

# Pinned versions with wheels available on Streamlit Cloud (py311 manylinux)
streamlit==1.36.0
playwright==1.46.0
pandas==2.2.2
python-dotenv==1.0.1
openai==1.40.3
tenacity==8.2.3
httpx==0.27.2
pydantic==2.8.2
PyMuPDF==1.24.9
SQLAlchemy==2.0.29
greenlet==2.0.2
supabase==2.4.0
